refactor+docs: move dev dialogs into ui package + update settings readme
