+++ app/src/dev/java/xyz/lbres/trickcalculator/ProductFlavor.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.trickcalculator
 
-import xyz.lbres.trickcalculator.devtools.DeveloperToolsDialog
+import xyz.lbres.trickcalculator.ui.devtools.DeveloperToolsDialog
 import xyz.lbres.trickcalculator.utils.visible
 
 /**
